examples/sqla_column_editable/pyproject.toml
[project]
name = "example-sqla-column-editable"
version = "0.1.0"
description = "SQLAlchemy column_editable_list example for Flask-Admin."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask-admin[sqlalchemy-with-utils,translation]",
]

[tool.uv.sources]
flask-admin = { path = "../../", editable = true }
